refactored to standards created more helpers
