--- subgraphs/compound-v2/src/common/utils.ts
@@ -1,1 +1,0 @@
-// store common calculations
