+++ subgraphs/compound-v2/src/common/utils/utils.ts
@@ -1,0 +1,11 @@
+// store common calculations
+import { BigDecimal } from "@graphprotocol/graph-ts";
+
+// turn exponent into a BigDecimal number
+export function exponentToBigDecimal(decimals: i32): BigDecimal {
+  let bigDecimal = BigDecimal.fromString("1");
+  for (let i = 0; i < decimals; i++) {
+    bigDecimal = bigDecimal.times(BigDecimal.fromString("10"));
+  }
+  return bigDecimal;
+}
