+++ subgraphs/compound-v2/src/mappings/mappings.ts
@@ -10,12 +10,11 @@ import {
   SUBGRAPH_VERSION,
   SCHEMA_VERSION,
   BIGDECIMAL_ZERO,
-  ZERO_ADDRESS,
-} from "../common/constants";
+} from "../common/utils/constants";
 
-import { createMarket, getAmountUSD } from "./helpers";
+import { createBorrow, createDeposit, createLiquidation, createMarket, createRepay, createWithdraw } from "./helpers";
 
-import { Mint, Redeem, Borrow as BorrowEvent, RepayBorrow, LiquidateBorrow } from "../types/cCOMP/cToken";
+import { Mint, Redeem, Borrow, RepayBorrow, LiquidateBorrow, Transfer } from "../types/templates/cToken/CToken";
 
 import {
   MarketEntered,
@@ -27,215 +26,45 @@ import {
   NewPriceOracle,
 } from "../types/Comptroller/Comptroller";
 
-import {
-  LendingProtocol,
-  Market,
-  Deposit,
-  Withdraw,
-  Borrow as BorrowEntity,
-  Repay,
-  Liquidation,
-  Token,
-} from "../types/schema";
-
-import { COMPTROLLER_ADDRESS } from "../common/addresses";
+import { LendingProtocol } from "../types/schema";
 import { CToken } from "../types/templates";
 import { AccrueInterest, NewMarketInterestRateModel, NewReserveFactor } from "../types/Comptroller/cToken";
-import { Transfer } from "../types/templates/CToken/CToken";
-import { Address, log } from "@graphprotocol/graph-ts";
 
 export function handleMint(event: Mint): void {
-  // get reused vars
-  let marketAddress = event.transaction.to!; // TODO: check if exists (if not, transaction doesn't exist)
-  let blockNumber = event.block.number;
-  let transactionHash = event.transaction.hash.toHexString();
-  let logIndex = event.logIndex;
-  let id = transactionHash + "-" + logIndex.toString();
-
-  // create new Deposit
-  let deposit = new Deposit(id);
-
-  // fill in deposit vars
-  deposit.hash = transactionHash;
-  deposit.logIndex = logIndex.toI32();
-  deposit.protocol = COMPTROLLER_ADDRESS.toHexString();
-  deposit.to = marketAddress.toHexString();
-  deposit.from = event.params.minter.toHexString();
-  deposit.blockNumber = blockNumber;
-  deposit.timestamp = event.block.timestamp;
-  deposit.market = marketAddress.toHexString();
-  let market = Market.load(marketAddress.toHexString());
-  if (market == null) {
-    // TODO: don't create deposit if this is the case
-    deposit.asset = ZERO_ADDRESS;
-    log.error("Market {} does not exist", [marketAddress.toHexString()]);
-    return;
-  } else {
-    deposit.asset = market.inputTokens[0];
+  if (createDeposit(event, event.params.mintAmount, event.params.minter)) {
+    // TODO: continue to handle usage metrics, financials, and market updates
   }
-  deposit.amount = event.params.mintAmount; // TODO: change to actual amount not 8000000...00
-  deposit.amountUSD = getAmountUSD(market, deposit.amount, blockNumber.toI32());
-
-  deposit.save();
 }
 
 export function handleRedeem(event: Redeem): void {
-  // reused vars
-  let marketAddress = event.transaction.from;
-  let blockNumber = event.block.number;
-  let transactionHash = event.transaction.hash.toHexString();
-  let logIndex = event.logIndex;
-  let id = transactionHash + "-" + logIndex.toString();
-
-  // creates Withdraw entity
-  let withdraw = new Withdraw(id);
-
-  // fill in withdraw vars
-  withdraw.hash = transactionHash;
-  withdraw.logIndex = logIndex.toI32();
-  withdraw.protocol = COMPTROLLER_ADDRESS.toHexString();
-  let toAddress = event.transaction.to;
-  if (toAddress) {
-    // TODO: these transactions aren't going through
-    withdraw.to = toAddress.toHexString();
-  } else {
-    withdraw.to = ZERO_ADDRESS;
+  if (createWithdraw(event, event.params.redeemer, event.params.redeemAmount)) {
+    // TODO: more things to update
   }
-  withdraw.from = marketAddress.toHexString();
-  withdraw.blockNumber = blockNumber;
-  withdraw.timestamp = event.block.timestamp;
-  withdraw.market = marketAddress.toHexString();
-  let market = Market.load(marketAddress.toHexString());
-  if (market == null) {
-    withdraw.asset = ZERO_ADDRESS;
-    log.error("Market {} does not exist", [marketAddress.toHexString()]);
-    return;
-  } else {
-    withdraw.asset = market.inputTokens[0];
-  }
-  withdraw.amount = event.params.redeemAmount;
-  withdraw.amountUSD = getAmountUSD(market, withdraw.amount, blockNumber.toI32());
-
-  withdraw.save();
 }
 
-export function handleBorrow(event: BorrowEvent): void {
-  // reused vars
-  let marketAddress = event.transaction.from;
-  let blockNumber = event.block.number;
-  let transactionHash = event.transaction.hash.toHexString();
-  let logIndex = event.logIndex;
-  let id = transactionHash + "-" + logIndex.toString();
-
-  // creates Borrow entity
-  let borrow = new BorrowEntity(id);
-
-  // fill in borrow vars
-  borrow.hash = transactionHash;
-  borrow.logIndex = logIndex.toI32();
-  borrow.protocol = COMPTROLLER_ADDRESS.toHexString();
-  borrow.to = event.params.borrower.toHexString();
-  borrow.from = marketAddress.toHexString();
-  borrow.blockNumber = blockNumber;
-  borrow.timestamp = event.block.timestamp;
-  borrow.market = marketAddress.toHexString();
-  let market = Market.load(marketAddress.toHexString());
-  if (market == null) {
-    borrow.asset = ZERO_ADDRESS;
-    log.error("Market {} does not exist", [marketAddress.toHexString()]);
-    return;
-  } else {
-    borrow.asset = market.inputTokens[0];
-  }
-  borrow.amount = event.params.borrowAmount;
-  borrow.amountUSD = getAmountUSD(market, borrow.amount, blockNumber.toI32());
-
-  borrow.save();
-
-  // update Market canBorrowFrom marketAddress
-  if (market != null) {
-    market.canBorrowFrom = true;
-    market.save();
+export function handleBorrow(event: Borrow): void {
+  if (createBorrow(event, event.params.borrower, event.params.borrowAmount)) {
+    // TODO: more updates
   }
 }
 
 export function handleRepayBorrow(event: RepayBorrow): void {
-  // reused vars
-  let marketAddress = event.transaction.to!; // TODO: check on this (if null transaction doesn't exist)
-  let blockNumber = event.block.number;
-  let transactionHash = event.transaction.hash.toHexString();
-  let logIndex = event.logIndex;
-  let id = transactionHash + "-" + logIndex.toString();
-
-  // create Repay entity
-  let repay = new Repay(id);
-
-  // populate repay vars
-  repay.hash = transactionHash;
-  repay.logIndex = logIndex.toI32();
-  repay.protocol = COMPTROLLER_ADDRESS.toHexString();
-  repay.to = marketAddress.toHexString();
-  repay.from = event.params.payer.toHexString();
-  repay.blockNumber = blockNumber;
-  repay.timestamp = event.block.timestamp;
-  repay.market = marketAddress.toHexString();
-  let market = Market.load(marketAddress.toHexString());
-  if (market == null) {
-    repay.asset = ZERO_ADDRESS;
-    log.error("Market {} does not exist", [marketAddress.toHexString()]);
-    return;
-  } else {
-    repay.asset = market.inputTokens[0];
+  if (createRepay(event, event.params.payer, event.params.repayAmount)) {
+    // TODO: more updates
   }
-  repay.amount = event.params.repayAmount;
-  repay.amountUSD = getAmountUSD(market, repay.amount, blockNumber.toI32());
-
-  repay.save();
 }
 
 export function handleLiquidateBorrow(event: LiquidateBorrow): void {
-  // TODO: verify market, to, from is accurate (changed)
-  // reused vars
-  let marketAddress = event.params.cTokenCollateral;
-  let blockNumber = event.block.number;
-  let transactionHash = event.transaction.hash.toHexString();
-  let logIndex = event.logIndex;
-  let id = transactionHash + "-" + logIndex.toString();
-
-  // create liquidation entity
-  let liquidation = new Liquidation(id);
-
-  // populate liquidations vars
-  liquidation.hash = transactionHash;
-  liquidation.logIndex = logIndex.toI32();
-  liquidation.protocol = COMPTROLLER_ADDRESS.toHexString();
-  liquidation.to = marketAddress.toHexString();
-  liquidation.from = event.params.liquidator.toHexString();
-  liquidation.blockNumber = blockNumber;
-  liquidation.timestamp = event.block.timestamp;
-  liquidation.market = marketAddress.toHexString();
-  let market = Market.load(marketAddress.toHexString());
-  if (market == null) {
-    liquidation.asset = ZERO_ADDRESS;
-    log.error("Market {} does not exist", [marketAddress.toHexString()]);
-    return;
-  } else {
-    liquidation.asset = market.inputTokens[0];
-  }
-  liquidation.amount = event.params.repayAmount;
-  liquidation.amountUSD = getAmountUSD(market, liquidation.amount, blockNumber.toI32());
-
-  // TODO: calculate profit
-  // liquidation.profitUSD = TODO
-
-  liquidation.save();
-
-  // update canUseAsCollateral in Market
-  let collateralToken = event.params.cTokenCollateral.toHexString();
-  let collateralMarket = Market.load(collateralToken);
-  if (collateralMarket != null) {
-    collateralMarket.canUseAsCollateral = true;
-    collateralMarket.save();
+  if (
+    createLiquidation(
+      event,
+      event.params.cTokenCollateral,
+      event.params.liquidator,
+      event.params.seizeTokens,
+      event.params.repayAmount,
+    )
+  ) {
+    // TODO: more updates
   }
 }
